+++ lib/compile/components/messages/index.js
@@ -1,24 +1,4 @@
-const { getEventName, resolveI18n } = require('../../utils')
-
-/**
- * Extracts description from event definition with fallback chain:
- * 1. @description annotation (with i18n resolution)
- * 2. @Core.Description annotation (with i18n resolution)
- * 3. doc comment (plain text)
- * @param {object} definition - Event definition from CSN
- * @param {object} i18nBundle - CDS i18n bundle from cds.i18n.bundle4(csn)
- * @return {string|undefined} Description text or undefined if none found
- */
-function getDescription(definition, i18nBundle) {
-    if (typeof definition['@description'] === 'string') {
-        return resolveI18n(definition['@description'], i18nBundle)
-    } else if (typeof definition['@Core.Description'] === 'string') {
-        return resolveI18n(definition['@Core.Description'], i18nBundle)
-    } else if (typeof definition.doc === 'string') {
-        return definition.doc.replace(/\n/g, ' ')
-    }
-    return undefined
-}
+const { getEventName, getDescription } = require('../../utils')
 
 const annotationMapping = {
     'EventSpecVersion': 'x-sap-event-spec-version',
