+++ lib/compile/components/schemas/annotations.js
@@ -1,6 +1,6 @@
 
 const cds = require('@sap/cds/lib')
-const { resolveI18n } = require('../../utils')
+const { resolveI18n, getDescription } = require('../../utils')
 
 const odmAnnotationsMapping = Object.freeze({
     '@ODM.entityName': 'x-sap-odm-entity-name',
@@ -20,12 +20,9 @@ module.exports = function addAnnotations(definition, object, i18nBundle) {
         object.title = resolveI18n(definition['@title'], i18nBundle)
     }
 
-    if (typeof (definition['@description']) === 'string') {
-        object.description = resolveI18n(definition['@description'], i18nBundle)
-    } else if (typeof (definition['@Core.Description']) === 'string') {
-        object.description = resolveI18n(definition['@Core.Description'], i18nBundle)
-    } else if (typeof definition.doc === 'string') {
-        object.description = definition.doc.replace(/\n/g, ' ')
+    const description = getDescription(definition, i18nBundle)
+    if (description) {
+        object.description = description
     }
 
     if (definition['@ODM.root']) {
