+++ lib/compile/utils.js
@@ -38,8 +38,29 @@ function resolveI18n(text, i18nBundle) {
     return i18nBundle?.at(match[1]) ?? text
 }
 
+/**
+ * Extracts description from definition with fallback chain:
+ * 1. @description annotation (with i18n resolution)
+ * 2. @Core.Description annotation (with i18n resolution)
+ * 3. doc comment (plain text)
+ * @param {object} definition - Event or element definition from CSN
+ * @param {object} i18nBundle - CDS i18n bundle from cds.i18n.bundle4(csn)
+ * @return {string|undefined} Description text or undefined if none found
+ */
+function getDescription(definition, i18nBundle) {
+    if (typeof definition['@description'] === 'string') {
+        return resolveI18n(definition['@description'], i18nBundle)
+    } else if (typeof definition['@Core.Description'] === 'string') {
+        return resolveI18n(definition['@Core.Description'], i18nBundle)
+    } else if (typeof definition.doc === 'string') {
+        return definition.doc.replace(/\n/g, ' ')
+    }
+    return undefined
+}
+
 module.exports = {
     getEventName,
     getOwnValue,
-    resolveI18n
+    resolveI18n,
+    getDescription
 }
